% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionetworks_logo.R
\name{bionetworks_logo}
\alias{bionetworks_logo}
\title{bionetworks_logo}
\usage{
bionetworks_logo()
}
\description{
Welcome to the bionetworks package!
}
\examples{
bionetworks_logo()

}
\author{
Xiaotao Shen
}
\keyword{internal}
