% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_core_nodes_degrees.R
\name{compare_core_nodes_degrees}
\alias{compare_core_nodes_degrees}
\title{Comparing the degrees between Rich-club and non-Rich-club genes}
\usage{
compare_core_nodes_degrees(study_network, rc_network)
}
\arguments{
\item{study_network}{a study network igraph object}

\item{rc_network}{a rich-club network igraph object}
}
\value{
Dataframe with degrees
}
\description{
Comparing the degrees between Rich-club and non-Rich-club genes
}
\examples{
compare_core_nodes_degrees(study_network, rc_network)
}
