% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_rc_coeff.R
\name{compare_rc_coeff}
\alias{compare_rc_coeff}
\title{Calculate rich club coefficient}
\usage{
compare_rc_coeff(study_net, rand_nets, weighted = F)
}
\arguments{
\item{weighted}{the option to calculate an edge weighted or unweighted rich-club coefficient}

\item{rand_networks}{a list of random networks}
}
\value{
Average rich club coefficient for a list of networks
}
\description{
This function calculates the rich club coefficient for the study network and compares it with
a list of random networks.
}
\examples{
rand_net_clu(rand_networks)

}
