% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_network.R
\name{construct_network}
\alias{construct_network}
\title{Construct interactive network}
\usage{
construct_network(network.df, interaction_score = "col_name")
}
\arguments{
\item{network.df}{a dataframe with interacting genes/proteins in columns 1 and 2. And an additional column of interaction strength (e.g. co-expression values)}

\item{interaction_score}{Name of column containing the interaction scores}
}
\value{
Returns as interactive network
}
\description{
This function constructs an interactive network using the pre-identified interactions and co-epxression values.
}
\examples{
construct_network(network.df, interaction_column ="Correlation_r")
}
