% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_rand_net.R
\name{construct_rand_net}
\alias{construct_rand_net}
\title{Construction of random networks}
\usage{
construct_rand_net(study_network, number = 100)
}
\arguments{
\item{network}{an igraph network object}
}
\value{
Visualize the degrees distribution in the interaction network
}
\description{
Construction of random networks of similar weighted degree distribution based on Fabien Viger and Matthieu Latapy algorithm.
}
\examples{
construct_rand_net(study_network, number = 100)

}
