% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_eigengenes.R
\name{core_eigengenes}
\alias{core_eigengenes}
\title{Calculate rich-club nodes eigengenes}
\usage{
core_eigengenes(data, study_network, rc_network)
}
\arguments{
\item{data}{microarray counts dataframe with rownames as genes and observations as columns}

\item{study_network}{a study network igraph object}

\item{rc_network}{a rich-club network igraph object}
}
\value{
Dataframe with degrees
}
\description{
Calculate rich-club nodes eigengenes
}
\examples{
core_eigengenes(study_network, rc_network)
}
