% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_nodes.R
\name{core_nodes}
\alias{core_nodes}
\title{Identify rich-club nodes in a network}
\usage{
core_nodes(study_network, rc_coeff.df, weighted = F)
}
\arguments{
\item{study_network}{a study network igraph object}

\item{rc_coeff.df}{a dataframe containing the calculated rich-club coefficients for the
study network and random networks, along with normalized values. Calculated using the compare_rc_coeff() function.}

\item{weighted}{the option to calculate an edge weighted or unweighted rich-club coefficient}
}
\value{
Network of hub nodes
}
\description{
This function identifies rich-club nodes (hubs) using the pre-computed rich-club coefficients for
the sutdy network and random networks
}
\examples{
core_nodes(study_network,rc_coeff.df,weighted=F)
}
