% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_nodes_correlation.R
\name{core_nodes_correlation}
\alias{core_nodes_correlation}
\title{Compare the correlation between Rich-club genes expression versus non-Rich-club genes with outcomes}
\usage{
core_nodes_correlation(
  rc_eigengenes.df,
  non_rc_eigengenes.df,
  outcomes.df,
  tested_outcome
)
}
\arguments{
\item{rc_eigengenes.df}{Cacluated using rc_eigengenes() function}

\item{non_rc_eigengenes.df}{Cacluated using non_rc_eigengenes() function}

\item{outcomes.df}{an outcomes dataframe for the same observations}

\item{tested_outcome}{outcome in the outcomes.df to be tested}
}
\value{
Predictions
}
\description{
Compare the correlation between Rich-club genes expression versus non-Rich-club genes with outcomes,
using a GLM model with PCA 1, 2 and 3.
}
\examples{
core_nodes_correlation <- function(rc_eigengenes.df, non_rc_eigengenes.df, outcomes.df, tested_outcome)
}
