% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degs_f_annotation.R
\name{degs_f_annotation}
\alias{degs_f_annotation}
\title{DEGs functional annotations}
\usage{
degs_f_annotation(genes, species = c("Hs", "Mm"))
}
\arguments{
\item{genes}{Gene symbols}

\item{control}{beginning label of control observations e.g. CTL (in: CTL.1, CTL.2,..)}

\item{condition}{beginning label of condition observations e.g. Rx (in: Rx.1, Rx.2,..)}
}
\value{
A dataframe with Gene Ontology
}
\description{
This function functionally annotates the identifed DEGs using Gene ontology analysis (biological processes, cellular components and molecular functions)
}
\examples{
degs_f_annotation(genes=genes, species = "Hs")
}
