% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_clusters.R
\name{genes_clusters}
\alias{genes_clusters}
\title{Genes co-expression using WGCNA}
\usage{
genes_clusters(data = data, deg = deg)
}
\arguments{
\item{data}{Expression counts table}

\item{deg}{A vector of differentially expressed genes}

\item{method}{The method to use to calculate the co-expression values}
}
\value{
A dataframe with gene-gene co-expression score
}
\description{
This function calculated genes co-expression using WGCNA soft-power adjusted adjacency matrix
}
\examples{
genes_clusters(data = data, method = c("pearson", "scaled"))
}
