% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_coexpression_s.R
\name{genes_coexpression}
\alias{genes_coexpression}
\title{Genes co-expression using WGCNA}
\usage{
genes_coexpression(data = data, deg = deg, method = c("pearson", "scaled"))
}
\arguments{
\item{data}{microarray counts table}
}
\value{
A dataframe with gene-gene co-expression score
}
\description{
This function calculated genes co-expression using WGCNA soft-power adjusted adjacency matrix
}
\examples{
genes_coexpression(data = data, method = c("pearson", "scaled"))
}
