% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_coexpression_r.R
\name{genes_coexpression_r}
\alias{genes_coexpression_r}
\title{Genes co-expression using Pearson correlation}
\usage{
genes_coexpression_r(data = data)
}
\arguments{
\item{data}{microarray counts table}
}
\value{
A dataframe with gene-gene co-expression pearson's r
}
\description{
This function calculated genes co-expression using pearson's correlation r value
}
\examples{
genes_coexpression_r(data = data)
}
