% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ident_interactions.R
\name{ident_interactions}
\alias{ident_interactions}
\title{Identify interactions}
\usage{
ident_interactions(genes_list, species = c("Hs", "Mm"), score_threshold = 200)
}
\arguments{
\item{data}{microarray counts table}
}
\value{
A dataframe with gene-gene interactions
}
\description{
This function constructs a gene/protein interactive network using interactions from STRING-database
}
\examples{
ident_interactions(genes_list, species = "Hs", score_threshold = 200)
}
