% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_net_construction.R
\name{int_net_construction}
\alias{int_net_construction}
\title{Interaction network construction}
\usage{
int_net_construction(
  genes_list,
  species = c("Hs", "Mm"),
  score_threshold = 200
)
}
\arguments{
\item{data}{microarray counts table}
}
\value{
A dataframe with gene-gene interactions
}
\description{
This function constructs a gene/protein interactive network using interactions from STRING-database
}
\examples{
int_net_construction(genes_list,species= "Hs", score_threshold=200)
}
