% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_int_expr.R
\name{merge_int_expr}
\alias{merge_int_expr}
\title{Merge gene interaction and co-expression data}
\usage{
merge_int_expr(
  gene_int.df,
  gene.coexpr.df,
  int_cols = c("int.col1", "int.col2"),
  coexpr_cols = c("exp.col1", "exp.col2")
)
}
\arguments{
\item{gene_int.df}{dataframe with genes interactions}

\item{int_cols}{names of the two gene names columns in the gene_int.df}

\item{gene_coexpr}{dataframe with genes co-expression score}

\item{expr_cols}{names of the two gene names columns in the gene_coexpr.df}
}
\value{
A dataframe with gene-gene co-expression score
}
\description{
This function merges gene interaction and co-expression data.
Function will adjust case sensitivity as needed.
}
\examples{
merge_int_expr(gene_int.df,gene.coexpr.df, int_cols= c("gene1","gene2"),
 expr_cols=c("Var1","Var2"))
}
