% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray_DEG.R
\name{microarray_degs}
\alias{microarray_degs}
\title{Microarray DEGs}
\usage{
microarray_degs(data, control, condition)
}
\arguments{
\item{data}{Microarray counts table with observations as columns and genes as rows with gene names as row-names}

\item{control}{beginning label of control observations e.g. CTL (in: CTL.1, CTL.2,..)}

\item{condition}{beginning label of condition observations e.g. Rx (in: Rx.1, Rx.2,..)}
}
\value{
Returns names of DEGs
}
\description{
This function identifies differentially expressed genes (DEGs) from microarray data
}
\examples{
microarray_DEGS(data= data, control ="CTL",condition= "Rx")
}
