% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_core_eigengenes.R
\name{non_core_eigengenes}
\alias{non_core_eigengenes}
\title{Calculate non-rich-club nodes eigengenes}
\usage{
non_core_eigengenes(data, study.net, core.nodes, number = 100)
}
\arguments{
\item{data}{microarray counts dataframe with rownames as genes and observations as columns}

\item{study.net}{a study network igraph object}

\item{core.nodes}{a rich-club network igraph object}

\item{number}{number of sampling iterations to perform}
}
\value{
Dataframe with degrees
}
\description{
Calculate non-rich-club nodes eigengenes for multiple different sampling of network node sets with
similar number to the rich-club nodes
}
\examples{
non_core_eigengenes(study_network, rc_network)
}
