% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_degrees_dist.R
\name{plot_degrees_dist}
\alias{plot_degrees_dist}
\title{Plot network degrees distribution}
\usage{
plot_degrees_dist(network)
}
\arguments{
\item{network}{an igraph network object}
}
\value{
Visualize the degrees distribution in the interaction network
}
\description{
This function visualizes the degrees distribution in the interaction network,
to ensure a scale-free network topology with power-law degrees distribution.
}
\examples{
plot_degrees_dist(network)
}
