% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_nets_rc.R
\name{rand_nets_rc}
\alias{rand_nets_rc}
\title{Calculate average rich club coefficient for a list of networks}
\usage{
rand_nets_rc(rand_nets, weighted = F)
}
\arguments{
\item{weighted}{the option to calculate an edge weighted or unweighted rich-club coefficient}

\item{rand_networks}{a list of random networks}
}
\value{
Average rich club coefficient for a list of networks
}
\description{
Calculate average rich club coefficient for a list of networks
}
\examples{
rand_net_clu(rand_networks)

}
