% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_eigengenes.R
\name{rc_eigengenes}
\alias{rc_eigengenes}
\title{Calculate rich-club nodes eigengenes}
\usage{
rc_eigengenes(data, study_network, rc_network)
}
\arguments{
\item{data}{microarray counts dataframe with rownames as genes and observations as columns}

\item{study_network}{a study network igraph object}

\item{rc_network}{a rich-club network igraph object}
}
\value{
Dataframe with degrees
}
\description{
Calculate rich-club nodes eigengenes
}
\examples{
compare_rc_degrees(study_network, rc_network)
}
