% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_net.R
\name{visualize_net}
\alias{visualize_net}
\title{Visualize interactive network}
\usage{
visualize_net(network)
}
\arguments{
\item{network}{an igraph network object}
}
\value{
Returns as interactive network visualized on Cytoscape
}
\description{
This function visualizes the interaction network with Cytoscape.
Must have Cytoscape application open before using this function.
}
\examples{
visualize_net(network)
}
