% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_coeff.R
\name{rc_coeff}
\alias{rc_coeff}
\title{rc_coeff}
\usage{
rc_coeff(g, k = 1, weighted = FALSE, A = NULL)
}
\arguments{
\item{study_network}{a study network igraph object}

\item{rc_network}{a rich-club network igraph object}
}
\value{
Dataframe with degrees
}
\description{
Comparing the degrees between Rich-club and non-Rich-club genes
}
\examples{
compare_core_nodes_degrees(study_network, rc_network)
}
